mocked stock search and history screen

+++ android/app/src/main/java/com/example/investmentportfolio/data/OperationItem.kt
@@ -1,3 +1,3 @@
 package com.example.investmentportfolio.data
 
-data class OperationItem(val isSale: Boolean, val date: String, val price: Int)
+data class OperationItem(val isSale: Boolean, val date: String, val price: Double)
